Added placeholder block, removed error message when breaking air, added direction based wood (texture not rotated), added int constants for chunk loader type, corrected places where the seed was treated as a long, added generateLevelData, writeToNbt and readFromNbt methods to WorldGeneratorBase, added some helper methods for writing BlockPos in NbtHelper, moved EnumAxis to its own file, added directionId to direction, marked several classes in Util as static, added a way to draw a Bounds in DebugDrawer, added ++ and -- operator overrides to BlockPos, added a Bounds field to Chunk, renamed SaveHelper to NbtIOHelper, fixed TileEntityGameObject gameObjects not being deleted when a wolrd is unloaded, moved StoneLayer out of WorldData, added an AbsVec helper method, added Mineshafts.

+++ Assets/VoxelEngine/Generation/CellularAutomaton/WorldGeneratorCellularAutomaton.cs
@@ -9,7 +9,7 @@ namespace VoxelEngine.Generation.CellularAutomaton {
         public GeneratorOptions options;
         private CaveGenerator caveGenerator;
 
-        public WorldGeneratorCellularAutomaton(World world, long seed) : base(world, seed) {
+        public WorldGeneratorCellularAutomaton(World world, int seed) : base(world, seed) {
             this.options = GeneratorOptions.DEFAULT_OPTIONS;
 
             this.caveGenerator = new CaveGenerator();
